intro_panel organizations cleanup, collect and rename key_concept_panels
