+++ client/src/panels/intro_panels/intro_panel_text_provider.js
@@ -1,13 +1,8 @@
-import intro_lang from './intro_lang.yaml';
-import fin_lang from "./financial_intro.yaml";
-import ppl_lang from "./people_intro.yaml";
-import results_lang from "./results_intro.yaml";
-import tag_lang from "./tagging_intro.yaml";
-import { 
-  util_components,
-} from '../shared.js';
+import intro_lang from "./intro_lang.yaml";
+import simplographic from "./simplographic.yaml";
+import gov_related from "./gov_related.yaml";
 
-const { create_text_maker_component, TrivialTextMaker} = util_components;
+import { util_components } from '../shared.js';
+const { create_text_maker_component } = util_components;
 
-export const { text_maker, TM } = create_text_maker_component([intro_lang, fin_lang, ppl_lang, results_lang, tag_lang]);
-export const TextMaker = TrivialTextMaker;
+export const { text_maker, TM } = create_text_maker_component([intro_lang, simplographic, gov_related]);
