+++ client/src/panels/intro_panels/key_concept_panels/people_questions.yaml
